# docs/swagger.yaml
basePath: /api/v1
definitions:
  fiber.Map:
    additionalProperties: true
    type: object
  github_com_mr-kaynak_go-core_internal_core_errors.ErrorCode:
    enum:
    - AUTH_001
    - AUTH_002
    - AUTH_003
    - AUTH_004
    - AUTH_005
    - AUTH_006
    - AUTH_007
    - USER_001
    - USER_002
    - USER_003
    - USER_004
    - VAL_001
    - VAL_002
    - VAL_003
    - VAL_004
    - DB_001
    - DB_002
    - DB_003
    - DB_004
    - DB_005
    - EXT_001
    - EXT_002
    - EXT_003
    - EXT_004
    - EMAIL_001
    - EMAIL_002
    - MQ_001
    - MQ_002
    - INTERNAL_001
    - NOT_FOUND
    - METHOD_NOT_ALLOWED
    - CONFLICT
    - BAD_REQUEST
    - UNPROCESSABLE_ENTITY
    type: string
    x-enum-varnames:
    - CodeUnauthorized
    - CodeInvalidCredentials
    - CodeTokenExpired
    - CodeTokenInvalid
    - CodeInsufficientRights
    - CodeAccountLocked
    - CodeAccountNotActivated
    - CodeUserNotFound
    - CodeUserAlreadyExists
    - CodeInvalidUserData
    - CodeEmailNotVerified
    - CodeValidationFailed
    - CodeInvalidInput
    - CodeMissingField
    - CodeInvalidFormat
    - CodeDatabaseError
    - CodeRecordNotFound
    - CodeDuplicateEntry
    - CodeConstraintViolated
    - CodeTransactionFailed
    - CodeServiceUnavailable
    - CodeTimeout
    - CodeRateLimitExceeded
    - CodeCircuitBreakerOpen
    - CodeEmailSendFailed
    - CodeInvalidEmail
    - CodeMessagePublishFailed
    - CodeMessageConsumeFailed
    - CodeInternal
    - CodeNotFound
    - CodeMethodNotAllowed
    - CodeConflict
    - CodeBadRequest
    - CodeUnprocessableEntity
  github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail:
    properties:
      code:
        $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ErrorCode'
      detail:
        type: string
      instance:
        type: string
      meta:
        additionalProperties: true
        type: object
      status:
        type: integer
      time:
        type: string
      title:
        type: string
      trace_id:
        type: string
      type:
        type: string
    type: object
...
  github_com_mr-kaynak_go-core_internal_modules_identity_domain.Metadata:
    additionalProperties: true
    type: object
  github_com_mr-kaynak_go-core_internal_modules_identity_domain.Permission:
    properties:
      category:
        description: e.g., "user", "role", "dashboard"
        type: string
      created_at:
        type: string
      description:
        description: Human-readable description
        type: string
      id:
        type: string
      name:
        description: e.g., "user.create", "user.delete", "role.manage"
        type: string
      updated_at:
        type: string
    type: object
  github_com_mr-kaynak_go-core_internal_modules_identity_domain.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_domain.Permission'
        type: array
      updated_at:
        type: string
    type: object
  github_com_mr-kaynak_go-core_internal_modules_identity_domain.User:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_verified:
        type: boolean
      last_login:
        type: string
      last_login_at:
        type: string
      last_name:
        type: string
      locked_until:
        type: string
      metadata:
        $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_domain.Metadata'
      phone:
        type: string
      roles:
        items:
          $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_domain.Role'
        type: array
      status:
        $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_domain.UserStatus'
      two_factor_enabled:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
      verified:
        type: boolean
    type: object
  github_com_mr-kaynak_go-core_internal_modules_identity_domain.UserStatus:
    enum:
    - active
    - inactive
    - locked
    - pending
    type: string
    x-enum-varnames:
    - UserStatusActive
    - UserStatusInactive
    - UserStatusLocked
    - UserStatusPending
...
  github_com_mr-kaynak_go-core_internal_modules_identity_service.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  github_com_mr-kaynak_go-core_internal_modules_identity_service.LoginResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_domain.User'
    type: object
  github_com_mr-kaynak_go-core_internal_modules_identity_service.RegisterRequest:
    properties:
      email:
        type: string
      first_name:
        maxLength: 50
        type: string
      last_name:
        maxLength: 50
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  github_com_mr-kaynak_go-core_internal_modules_identity_service.TokenPair:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
    type: object
...
  internal_modules_identity_api.EmailRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_modules_identity_api.LogoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
...
  internal_modules_identity_api.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  internal_modules_identity_api.ResetPasswordRequest:
    properties:
      new_password:
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
...
host: localhost:3000
info:
  contact:
    name: Go-Core Team
    url: https://github.com/mr-kaynak/go-core
  description: Enterprise-grade Go boilerplate with REST API, gRPC services, and notification
    system
  license:
    name: MIT
  title: Go-Core API
  version: 1.0.0
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_service.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_service.LoginResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate the current session and refresh token
      parameters:
      - description: Optional refresh token to revoke
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_modules_identity_api.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Not authenticated
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      security:
      - Bearer: []
      summary: Logout user
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: New token pair
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_service.TokenPair'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      summary: Refresh access token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user account with email verification
      parameters:
      - description: Registration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_modules_identity_service.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Registration successful
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      summary: Register a new user
      tags:
      - Auth
  /auth/request-password-reset:
    post:
      consumes:
      - application/json
      description: Send a password reset email
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.EmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reset email sent
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      summary: Request password reset
      tags:
      - Auth
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Resend the email verification link
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.EmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Verification email sent
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      summary: Resend verification email
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset password using a reset token
      parameters:
      - description: Reset token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_identity_api.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successful
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      summary: Reset password
      tags:
      - Auth
  /auth/validate-reset-token:
    get:
      description: Check if a password reset token is valid
      parameters:
      - description: Password reset token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token is valid
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      summary: Validate password reset token
      tags:
      - Auth
  /auth/verify-email:
    get:
      description: Verify user email with token from verification email
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Email verified successfully
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid or expired token
          schema:
            $ref: '#/definitions/github_com_mr-kaynak_go-core_internal_core_errors.ProblemDetail'
      summary: Verify email address
      tags:
      - Auth
...
securityDefinitions:
  Bearer:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"